agentes/agente-shopee/requirements.txt
# orquestrador/requirements.txt
flask
google-generativeai
# Dependências específicas para raspagem de dados
requests
beautifulsoup4